Document Integration Testing Using Subdomain Constraints (#8)
The README describes how to set up an API using a subdomain constraint,
so the API would be available at `api.yourdomain.com`. However, this
causes all integration tests in the API to fail because the subdomain is
not being used when looking up routes to send requests to. To remedy
this, a note has been added to the README explaining how to configure
`Workarea::IntegrationTest` so that subdomains can be used in the API
without integration tests suddenly failing.

**More Info:** https://discourse.workarea.com/t/build-failing-workarea-api-plugin/1652

No changelog