Fix Docs Generator Task Name
The task name `workarea:api:generate_docs` wasn't able to load the
environment like the rest of the `workarea:test` tasks because it wasn't
in the correct Rake namespace. This is due to a hack in core that
configures the Rake environment such that custom Rake tasks for testing
are possible. Change the task name to `workarea:test:api_docs` to take
advantage of this hack so the tests will run properly.

API-4