Fix Caching in JBuilder Templates
To resolve an issue where cached data would still appear in Jbuilder
templates even when requesting as a logged-in admin, override the
`JbuilderTemplate#cache!` method to perform the same checks as the
`ActionView::Helpers::WorkareaCache` helper when rendering pages as
HTML. This ensures that data between API requests and regular HTML
requests on the storefront are consistent. The patch is only active for
Storefront API controllers, as Jbuilder is used extensively elsewhere
and this kind of cache lookup wouldn't work in those contexts.

API-10