Fix Checkout Update Response
Consumers of the Checkouts API expect a 422 response when a failure
occurs for any reason. Previously, Workarea was always returnin ga 200
success even when the update could not be fully completed. To resolve
this, the `Workarea::Checkout#update` method has been modified to
return `false` when the update fails, triggering the proper status code
in the response.

API-13