Add Documentation Test for Failed Checkout Update
This documents the scenario in which a checkout update fails to persist,
and thus the API returns a `422` error to signify its failure. It
depends on the code changes in workarea-commerce/workarea#481, and thus
we will need to bump the Workarea version to restrict its upgrade to
users of v3.5.17 and beyond.

API-14